refactor(3124): keep every over-threshold file at or below its baseline
Merge the batched stream into the generic joined_scan_stream module (net-zero
module/re-export churn in data_access/mod.rs + reader/mod.rs), host the forwarder
under scan_stream_windowed, and split the per-row streaming scan out of
sequential.rs (1683 -> 1486, epic #1116). Also gate the fan-out-only imports and
the FanoutMerge site variant for tombstones builds (--all-features clippy).
