+++ cqlite-core/src/storage/sstable/reader/data_access/joined_scan_stream.rs
@@ -28,11 +28,50 @@
 //! `async fn(&mut self) -> Option<Result<T>>` signature `mpsc::Receiver::recv` had,
 //! so every `while let Some(item) = stream.recv().await` loop keeps working — and
 //! now fails closed.
+//!
+//! # The boundary #3106 closed first (the batched single-source scan)
+//!
+//! `scan_stream_batched_admitted` (`sequential.rs`) drives the scan in a spawned
+//! `tokio` task that owns the sending half of a bounded channel and reports a
+//! failure by sending `Err(..)`. Before this type it returned the bare
+//! `mpsc::Receiver` and DISCARDED the task's `JoinHandle`, so a task that
+//! UNWOUND — a decode panic in `parse_block_entries_at_now`, which on the
+//! non-stitching (uncompressed) branch runs directly inside that task, or any
+//! future panic on the walk — dropped its sender with no error and no terminator.
+//! Every consumer read that close as "the scan finished":
+//!
+//! * `query_rows::drive_full_scan_rows` returned `Ok(())`, its own thread then
+//!   sent the #3106 `Done` sentinel, and the Flight `do_get` completed
+//!   SUCCESSFULLY with a truncated result set — the #3106 defect exactly, one
+//!   layer below the channel #3106 first fixed, on the arm a `do_get` with no
+//!   token filter takes (i.e. the reported repro);
+//! * the query engine's streaming SELECT / aggregate folds
+//!   (`select_executor::streaming`, `stream_agg`, `executor::streaming_scan_rows`)
+//!   returned short row sets / wrong aggregates the same way.
+//!
+//! [`BatchedScanStream`] owns the handle instead, so a channel close is
+//! DISAMBIGUATED: the task is joined and a `JoinError` (panic or cancellation)
+//! surfaces as `Some(Err(..))` rather than `None`. This mirrors what the windowed
+//! sub-path already did internally for its own two child tasks
+//! (`scan_stream_windowed`'s `parse_task.await` / feed `join_err` arms map a
+//! `JoinError` to an error), so the two arms no longer diverge on whether a dead
+//! producer is an error.
+//!
+//! # The #3124 boundaries (the multi-generation path)
+//!
+//! `SSTableManager::scan_stream`'s fan-out k-way merge task, each per-generation
+//! `scan_stream_admitted` sub-scan, and the per-row → batch re-chunker's source had
+//! the identical shape on the ≠1-generation path, and the windowed forwarder
+//! discarded its `JoinError` outright. Their wiring lives in
+//! `storage/sstable/scan_stream_fanout.rs` and
+//! `storage/sstable/reader/scan_stream_forwarder.rs`; the mechanism they all use is
+//! this type.
 
 use tokio::sync::mpsc;
 use tokio::task::JoinHandle;
 
-use crate::types::ScanRow;
+use super::super::SSTableReader;
+use crate::types::{ScanRow, TableId};
 use crate::{Error, Result, RowKey};
 
 /// An item type a [`JoinedStream`] can carry, plus the human name of the surface
@@ -95,7 +134,10 @@ enum TaskState {
 
 impl<T: ScanStreamItem> JoinedStream<T> {
     /// Pair a scan channel with the task that drives it.
-    pub(in crate::storage::sstable) fn new(rx: mpsc::Receiver<Result<T>>, task: JoinHandle<()>) -> Self {
+    pub(in crate::storage::sstable) fn new(
+        rx: mpsc::Receiver<Result<T>>,
+        task: JoinHandle<()>,
+    ) -> Self {
         Self {
             rx,
             task: TaskState::Running(task),
@@ -192,8 +234,52 @@ fn dead_scan_task_error_after_report<T: ScanStreamItem>() -> Error {
 /// those boundaries rather than a short, successful-looking result set.
 pub type RowScanStream = JoinedStream<(RowKey, ScanRow)>;
 
+/// The batched streaming scan's consumer handle: the batch channel PLUS the driver
+/// task that feeds it, so end-of-stream is an observed fact (issue #3106). Behaviour
+/// and API are unchanged from when this was its own `struct` (issue #3106); only the
+/// item type is fixed here.
+pub type BatchedScanStream = JoinedStream<Vec<(RowKey, ScanRow)>>;
+
+impl SSTableReader {
+    /// Open the batched streaming scan's cursor — the FIRST thing the driver task
+    /// does, ABOVE the `requires_chunk_stitching()` branch, i.e. on every format.
+    ///
+    /// A named seam because it is the SINGLE test-only fault checkpoint for this
+    /// task (issue #3106), and it is here precisely because it is
+    /// format-branch-independent: a checkpoint inside either branch fires only for
+    /// the formats that take that branch, so it can silently not fire and leave a
+    /// test passing vacuously. Killing the task here reproduces exactly the
+    /// condition the fix is about — the task's sender drops with no error and no
+    /// terminator — for any reader, and the join that catches it wraps the whole
+    /// task, so the property proven is branch-agnostic.
+    pub(super) async fn open_batched_scan_cursor(&self) -> Result<super::ScanCursor> {
+        crate::storage::producer_fault::inner_scan_task_checkpoint(|| self.file_path());
+        self.new_scan_cursor().await
+    }
+
+    /// Decode one `Data.db` block for the batched NON-STITCHING streaming scan,
+    /// pinning `read_shadowing = true` for this scan in ONE place.
+    ///
+    /// Deliberately carries NO fault checkpoint (issue #3106, roborev): it is
+    /// reached only by readers whose format takes the non-stitching branch, and no
+    /// fixture in the tree does, so a checkpoint here would be an armable-but-never-
+    /// armed seam — latent confusion suggesting coverage that does not exist. The
+    /// task-level checkpoint in [`Self::open_batched_scan_cursor`] covers both
+    /// branches, which is what the join it proves is scoped to anyway.
+    pub(super) fn parse_batched_block(
+        &self,
+        block: &[u8],
+        schema: Option<&crate::schema::TableSchema>,
+        now_secs: Option<i64>,
+    ) -> Result<Vec<(TableId, RowKey, ScanRow)>> {
+        self.parse_block_entries_at_now(block, schema, true, now_secs)
+    }
+}
+
 #[cfg(test)]
 mod tests {
+    use tokio::sync::mpsc;
+
     use super::*;
 
     fn row(byte: u8) -> (RowKey, ScanRow) {
@@ -202,8 +288,8 @@ mod tests {
 
     /// Issue #3124: the PER-ROW flavour of the #3106 property — a producer task that
     /// DIED is reported as an error, and that verdict is sticky. The batched flavour
-    /// is pinned by the sibling tests in `batched_scan_stream.rs`; this proves the
-    /// generic machinery carries the property to the surface #3124 is about.
+    /// is pinned by the sibling tests below; this proves the generic machinery carries
+    /// the property to the per-row surface #3124 is about.
     #[tokio::test]
     async fn a_dead_per_row_task_is_reported_as_an_error_and_stays_one() {
         // Held across every await that can raise the injected panic; the guard
@@ -257,7 +343,215 @@ mod tests {
             let _ = tx.send(Ok(row(2))).await;
         });
         let mut stream = RowScanStream::new(rx, task);
-        assert!(matches!(stream.recv().await, Some(Ok(_))), "the row arrives");
+        assert!(
+            matches!(stream.recv().await, Some(Ok(_))),
+            "the row arrives"
+        );
+        assert!(
+            stream.recv().await.is_none(),
+            "a task that returned normally is a clean end of stream"
+        );
+        assert!(
+            stream.recv().await.is_none(),
+            "and stays one on a later poll"
+        );
+    }
+
+    /// Issue #3106: a driver task that DIED is reported as an error, and that
+    /// verdict is STICKY — a consumer that keeps polling can never coax a
+    /// "proven-clean end of stream" `None` out of a dead scan.
+    ///
+    /// The task is killed with a real panic (the same unwind a decode bug would
+    /// produce); its console noise is silenced by message so a genuine failure in
+    /// a parallel test is never masked.
+    #[tokio::test]
+    async fn a_dead_scan_task_is_reported_as_an_error_and_stays_one() {
+        // Held for the whole test: the guard filters ONLY the injected message and
+        // delegates every other panic to the previous hook, so an assertion failure
+        // below (here or in a parallel test) still prints. The task panics
+        // asynchronously, so a narrower scope would race the hook restore.
+        let _silence = crate::storage::producer_fault::silence_injected_panics();
+        let (tx, rx) = mpsc::channel::<Result<Vec<(RowKey, ScanRow)>>>(4);
+        let task = tokio::spawn(async move {
+            // Deliver one batch, then die WITHOUT reporting — the exact shape the
+            // discarded `JoinHandle` used to turn into a clean end of stream.
+            let _ = tx
+                .send(Ok(vec![(RowKey::new(vec![1]), ScanRow::Row(Vec::new()))]))
+                .await;
+            panic!("{}", crate::storage::producer_fault::INJECTED_PANIC_MESSAGE);
+        });
+        let mut stream = BatchedScanStream::new(rx, task);
+
+        let (first, rest) = {
+            let first = stream.recv().await;
+            let mut rest = Vec::new();
+            for _ in 0..3 {
+                rest.push(stream.recv().await);
+            }
+            (first, rest)
+        };
+        // The silencer (held above across every await that can raise the injected
+        // panic) is dropped here, BEFORE the assertions: restoring a panic hook
+        // from a panicking thread aborts the process, which would turn a
+        // legitimate failure below into an unreadable SIGABRT.
+        drop(_silence);
+        assert!(
+            matches!(first, Some(Ok(batch)) if batch.len() == 1),
+            "the batch produced before the task died is still delivered"
+        );
+        for (attempt, outcome) in rest.into_iter().enumerate() {
+            let msg = outcome
+                .expect("a dead task must never report a clean end of stream")
+                .expect_err("it must be an error")
+                .to_string();
+            assert!(
+                msg.contains("DIED without reporting") && msg.contains("TRUNCATED"),
+                "attempt {attempt}: the error must name the dead task and the \
+                 truncation, got: {msg}"
+            );
+        }
+    }
+
+    /// Build a stream whose task: sends one batch, CLOSES the channel, signals
+    /// that it has done so, then parks until released — so a `recv` that runs
+    /// after the close signal is guaranteed to be suspended in the JOIN, not in
+    /// the channel. `then_panic` decides how it finally exits.
+    ///
+    /// The close signal is what makes the cancellation test deterministic: without
+    /// it, a `recv` could be cancelled while still awaiting the channel, which
+    /// exercises nothing.
+    fn stream_parked_after_closing_the_channel(
+        then_panic: bool,
+    ) -> (
+        BatchedScanStream,
+        tokio::sync::oneshot::Receiver<()>,
+        tokio::sync::oneshot::Sender<()>,
+    ) {
+        let (tx, rx) = mpsc::channel::<Result<Vec<(RowKey, ScanRow)>>>(4);
+        let (closed_tx, closed_rx) = tokio::sync::oneshot::channel::<()>();
+        let (release_tx, release_rx) = tokio::sync::oneshot::channel::<()>();
+        let task = tokio::spawn(async move {
+            let _ = tx
+                .send(Ok(vec![(RowKey::new(vec![7]), ScanRow::Row(Vec::new()))]))
+                .await;
+            drop(tx);
+            let _ = closed_tx.send(());
+            let _ = release_rx.await;
+            if then_panic {
+                panic!("{}", crate::storage::producer_fault::INJECTED_PANIC_MESSAGE);
+            }
+        });
+        (BatchedScanStream::new(rx, task), closed_rx, release_tx)
+    }
+
+    /// Cancel a `recv` WHILE it is awaiting the join, then poll again.
+    ///
+    /// `biased` polls the `recv` first, so it runs to the join (the channel is
+    /// already closed and drained) and returns `Pending` because the task is
+    /// parked; the ready branch then completes the `select!` and DROPS the `recv`
+    /// future — a cancellation landing exactly on the join await.
+    async fn cancel_a_recv_mid_join(stream: &mut BatchedScanStream) {
+        tokio::select! {
+            biased;
+            _ = stream.recv() => panic!(
+                "test precondition: the join must still be pending — the driver \
+                 task is parked, so recv cannot have resolved"
+            ),
+            _ = std::future::ready(()) => {}
+        }
+    }
+
+    /// Issue #3106 (roborev blocker): a `recv` CANCELLED mid-join must not lose the
+    /// task's verdict. The pre-fix code moved the `JoinHandle` out of `self` before
+    /// awaiting it, so a cancellation dropped the handle (detaching the task) with
+    /// no verdict recorded — and the next `recv` short-circuited on the absent
+    /// handle to a clean end of stream for a task that then PANICKED. That is the
+    /// #3106 defect, re-entered through the door the fix built.
+    #[tokio::test]
+    async fn a_recv_cancelled_mid_join_still_reports_the_dead_task() {
+        let (mut stream, closed, release) = stream_parked_after_closing_the_channel(true);
+
+        assert!(
+            matches!(stream.recv().await, Some(Ok(batch)) if batch.len() == 1),
+            "the batch produced before the task parked is delivered"
+        );
+        closed
+            .await
+            .expect("the task signals that it closed the channel");
+        cancel_a_recv_mid_join(&mut stream).await;
+
+        // Now let the parked task die, and re-poll. The silencer is dropped BEFORE
+        // any assertion: restoring a panic hook from a panicking thread aborts the
+        // process, which would turn a legitimate failure below into an unreadable
+        // SIGABRT instead of a reported assertion.
+        let outcomes = {
+            let _silence = crate::storage::producer_fault::silence_injected_panics();
+            release.send(()).expect("release the parked task");
+            let mut outcomes = Vec::new();
+            for _ in 0..3 {
+                outcomes.push(stream.recv().await);
+            }
+            outcomes
+        };
+        for (attempt, outcome) in outcomes.into_iter().enumerate() {
+            let msg = outcome
+                .expect(
+                    "a cancelled join must NOT lose the verdict: reporting a clean \
+                     end of stream here is issue #3106 reintroduced",
+                )
+                .expect_err("the task panicked, so this must be an error")
+                .to_string();
+            assert!(
+                msg.contains("DIED without reporting") && msg.contains("TRUNCATED"),
+                "attempt {attempt}: the error must name the dead task and the \
+                 truncation, got: {msg}"
+            );
+        }
+    }
+
+    /// The complement: a cancelled join must not be turned into a spurious error
+    /// either. Same cancellation, but the task then finishes NORMALLY — the stream
+    /// must report the clean end of stream it genuinely observed.
+    #[tokio::test]
+    async fn a_recv_cancelled_mid_join_still_reports_a_clean_finish_as_clean() {
+        let (mut stream, closed, release) = stream_parked_after_closing_the_channel(false);
+
+        assert!(
+            matches!(stream.recv().await, Some(Ok(_))),
+            "the batch arrives"
+        );
+        closed
+            .await
+            .expect("the task signals that it closed the channel");
+        cancel_a_recv_mid_join(&mut stream).await;
+
+        release.send(()).expect("release the parked task");
+        assert!(
+            stream.recv().await.is_none(),
+            "a task that ran to completion is a clean end of stream, even though an \
+             earlier recv was cancelled mid-join"
+        );
+        assert!(
+            stream.recv().await.is_none(),
+            "and stays clean on a later poll"
+        );
+    }
+
+    /// The control: a task that finishes normally yields the clean end of stream,
+    /// so the fail-closed join above cannot be passing for a trivial reason.
+    #[tokio::test]
+    async fn a_finished_scan_task_yields_a_clean_end_of_stream() {
+        let (tx, rx) = mpsc::channel::<Result<Vec<(RowKey, ScanRow)>>>(4);
+        let task = tokio::spawn(async move {
+            let _ = tx
+                .send(Ok(vec![(RowKey::new(vec![2]), ScanRow::Row(Vec::new()))]))
+                .await;
+        });
+        let mut stream = BatchedScanStream::new(rx, task);
+        assert!(
+            matches!(stream.recv().await, Some(Ok(_))),
+            "the batch arrives"
+        );
         assert!(
             stream.recv().await.is_none(),
             "a task that returned normally is a clean end of stream"
